requirements.txt
# Project dependencies

dnspython

